move data fixtures to conftest
