+++ tests/conftest.py
@@ -3,9 +3,12 @@ from io import BytesIO
 from pathlib import Path
 from unittest.mock import Mock
 
+import pandas as pd
 import pytest
 import requests
 from response_cases import RESPONSE_CASES
+
+from datareservoirio._utils import DataHandler
 
 TEST_PATH = Path(__file__).parent
 
@@ -62,3 +65,43 @@ def bytesio_with_memory(monkeypatch):
 @pytest.fixture
 def auth_session():
     return requests.Session()
+
+
+@pytest.fixture
+def data_float():
+    """Data with float values"""
+    index_list = (
+        1640995215379000000,
+        1640995219176000000,
+        1640995227270000000,
+        1640995267223000000,
+        1640995271472000000,
+    )
+
+    values_list = (-0.2, -0.1, 0.2, 0.1, 1.2)
+
+    series = pd.Series(data=values_list, index=index_list, name="values")
+
+    data_handler = DataHandler(series)
+
+    return data_handler
+
+
+@pytest.fixture
+def data_string():
+    """Data with string values"""
+    index_list = (
+        1640995215379000000,
+        1640995219176000000,
+        1640995227270000000,
+        1640995267223000000,
+        1640995271472000000,
+    )
+
+    values_list = ("foo", "bar", "baz", "foobar", "abcd")
+
+    series = pd.Series(data=values_list, index=index_list, name="values")
+
+    data_handler = DataHandler(series)
+
+    return data_handler
