+++ tests/test_storage/test_cache_engine.py
@@ -11,41 +11,6 @@ TEST_PATH = Path(__file__).parent
 
 
 class Test_CacheIO:
-    @pytest.fixture
-    def data_float(self):
-        index_list = (
-            1640995215379000000,
-            1640995219176000000,
-            1640995227270000000,
-            1640995267223000000,
-            1640995271472000000,
-        )
-
-        values_list = (-0.2, -0.1, 0.2, 0.1, 1.2)
-
-        series = pd.Series(data=values_list, index=index_list, name="values")
-
-        data_handler = DataHandler(series)
-
-        return data_handler
-
-    @pytest.fixture
-    def data_string(self):
-        index_list = (
-            1640995215379000000,
-            1640995219176000000,
-            1640995227270000000,
-            1640995267223000000,
-            1640995271472000000,
-        )
-
-        values_list = ("foo", "bar", "baz", "foobar", "abcd")
-
-        series = pd.Series(data=values_list, index=index_list, name="values")
-
-        data_handler = DataHandler(series)
-
-        return data_handler
 
     @pytest.mark.parametrize("data", ("data_float", "data_string"))
     def test__write(self, request, data, tmp_path):
